% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_es.R
\name{cal_es}
\alias{cal_es}
\title{Calculate effect measures}
\usage{
cal_es(
  data,
  outcome,
  study,
  exposure,
  ref,
  measure = c("MD", "Cohen", "Glass", "CL", "Hedge", "Cliff", "RR", "RD", "OR"),
  hetsk = FALSE,
  design
)
}
\arguments{
\item{data}{name of the data frame}

\item{outcome}{column name of the outcome}

\item{exposure}{column name of the exposure}

\item{ref}{reference level of the exposure (default: 1st level)}

\item{measure}{six methods are available:
\itemize{
  \item "MD": Mean Difference;
  \item "Cohen": Cohan's D;
  \item "Glass": Glass's G;
  \item "CL": common language;
  \item "Hedge": Hedge's G;
  \item "Cliff": Cliff's \eqn{\delta}
  \item "RR" Risk Ratio;
  \item "RD" Risk Difference;
  \item "OR" Odds Ratio
}}

\item{hetsk}{indicator of heteroskedasticity}

\item{design}{design of 2x2 table}
}
\value{
a data frame of effect measures
}
\description{
Calculate effect measures
}
\examples{
\dontrun{
test
}
}
\author{
Marta Regis \email{m.regis@tue.nl}
}
