% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_het.R
\name{test_het}
\alias{test_het}
\title{Testing for heterogeneity}
\usage{
test_het(Yk, Sk, dfk, method = c("CQ", "Bliss", "CF"), hetsk = TRUE)
}
\arguments{
\item{Yk}{a vector of effect sizes}

\item{Sk}{a vector of standard errors of the effect sizes}

\item{dfk}{a vector of degrees of freedom}

\item{method}{three methods are available:
\itemize{
  \item "CQ": Cochran's Q-test;
  \item "Bliss": Bliss' \eqn{\chi^2}-test;
  \item "CF": Cochran's F-tests (w/o heteroskedasticity)
}}

\item{hetsk}{indicator of heteroskedasticity assumptions for Cochran's F-test method}
}
\value{
a list of test statistic, degrees of freedom, and p-value
}
\description{
\code{test_het} returns the test statistic according to the methods chosen
}
\examples{
test_het(yk, sk, dfk, method="CQ")
}
