% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_ES.R
\name{cal_ES}
\alias{cal_ES}
\title{Title}
\usage{
cal_ES(
  data,
  outcome,
  exposure,
  ref,
  measure = c("MD", "Cohen", "Glass", "CL", "Hedge", "Cliff"),
  hetsk = FALSE
)
}
\arguments{
\item{data}{name of the data frame}

\item{outcome}{column name of the outcome}

\item{exposure}{column name of the exposure}

\item{ref}{reference level of the exposure (default: 1st level)}

\item{measure}{six methods are available:
\itemize{
  \item "MD": mean difference;
  \item "Cohen": Cohan's D;
  \item "Glass": Glass's G;
  \item "CL": common language;
  \item "Hedge": Hedge's G;
  \item "Cliff": Cliff's \eqn{\delta}
}}

\item{hetsk}{indicator of heteroskedasticity}
}
\value{
a data frame of effect measures
}
\description{
Title
}
\author{
Marta Regis \email{m.regis@tue.nl}
}
